Added regression test
