+++ glue/core/data_factories/tests/test_data_factories.py
@@ -12,8 +12,7 @@ from glue.core.component import CategoricalComponent
 from glue.core.data import Data
 from glue.core import data_factories as df
 from glue.config import data_factory
-from glue.tests.helpers import (requires_astropy,
-                                requires_pil_or_skimage, make_file, requires_qt)
+from glue.tests.helpers import requires_astropy, make_file, requires_qt
 
 
 def test_load_data_auto_assigns_label():
@@ -40,28 +39,6 @@ def test_data_label():
     assert df.data_label('') == ''
     assert df.data_label('/Leading/Path/no_extension') == 'no_extension'
     assert df.data_label('no_extension') == 'no_extension'
-
-
-@requires_pil_or_skimage
-def test_grey_png_loader():
-    # Greyscale PNG
-    data = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x02\x00\x00\x00\x02\x08\x00\x00\x00\x00W\xddR\xf8\x00\x00\x00\x0eIDATx\x9ccdddab\x04\x00\x00&\x00\x0b\x8e`\xe7A\x00\x00\x00\x00IEND\xaeB`\x82'
-    with make_file(data, '.png') as fname:
-        d = df.load_data(fname)
-        assert df.find_factory(fname) is df.img_data
-    assert_array_equal(d['PRIMARY'], [[3, 4], [1, 2]])
-
-
-@requires_pil_or_skimage
-def test_color_png_loader():
-    # Colorscale PNG
-    data = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x02\x00\x00\x00\x02\x08\x02\x00\x00\x00\xfd\xd4\x9as\x00\x00\x00\x15IDAT\x08\xd7\x05\xc1\x01\x01\x00\x00\x00\x80\x10\xffO\x17B\x14\x1a!\xec\x04\xfc\xf2!Q\\\x00\x00\x00\x00IEND\xaeB`\x82'
-    with make_file(data, '.png') as fname:
-        d = df.load_data(fname)
-        assert df.find_factory(fname) is df.img_data
-    assert_array_equal(d['red'], [[255, 0], [255, 0]])
-    assert_array_equal(d['green'], [[255, 0], [0, 255]])
-    assert_array_equal(d['blue'], [[0, 255], [0, 0]])
 
 
 @pytest.mark.parametrize(('delim', 'suffix'),
